+++ test/generators.test.inc.js
@@ -37,6 +37,24 @@ describe('Generator', function() {
 		});
 	});
 
+	it('delays first iteration with option {immediate: false}', function() {
+		var order = [];
+		var fn = series(function*(i) {
+			order.push('start' + i);
+			yield Promise.resolve();
+			order.push('end' + i);
+			return i * 10;
+		}, {immediate: false});
+
+		var res = [1, 2, 3].map(fn);
+		order.push('sync');
+
+		return res[2].then(function(res) {
+			expect(res).to.equal(30);
+			expect(order).to.deep.equal(['sync', 'start1', 'end1', 'start2', 'end2', 'start3', 'end3']);
+		});
+	});
+
 	it('passes this context', function() {
 		var fn = series(function*(i) {
 			yield Promise.resolve();