+++ changelog.md
@@ -22,6 +22,7 @@
 ## Next
 
 * Execute first iteration immediately
+* `immediate` option
 * Only use `co.wrap` on generators
 
 Breaking changes: